Rename client JS files to kebab-case
