+++ src/javascripts/clipped-list.js
