+++ src/javascripts/create-investment.js
