+++ src/javascripts/expanding-cards.js
