+++ src/views/company/add-step-1.html
@@ -57,5 +57,5 @@
       <a href="/" class="button-link button--cancel">Cancel</a>
     </div>
   </form>
-  <script src="/javascripts/companyadd.bundle.js"></script>
+  <script src="/javascripts/company-add.bundle.js"></script>
 {% endblock %}
