+++ src/views/company/archive.html
@@ -26,5 +26,5 @@
       {{ trade.save(buttonText="Archive", backUrl=companyUrl) }}
     </form>
   </div>
-  <script src="/javascripts/archiveform.bundle.js"></script>
+  <script src="/javascripts/archive-form.bundle.js"></script>
 {% endif %}
