+++ src/views/company/contacts.html
@@ -81,5 +81,5 @@
 {% endif %}
 
 
-<script src="/javascripts/expandingcards.bundle.js"></script>
+<script src="/javascripts/expanding-cards.bundle.js"></script>
 {% endblock %}
