+++ src/views/company/edit-foreign.html
@@ -84,5 +84,5 @@
       </form>
     </div>
   </div>
-  <script src="/javascripts/companyedit.bundle.js"></script>
+  <script src="/javascripts/company-edit.bundle.js"></script>
 {% endblock %}
