+++ src/views/company/edit-ltd.html
@@ -107,5 +107,5 @@
       </form>
     </div>
   </div>
-  <script src="/javascripts/companyedit.bundle.js"></script>
+  <script src="/javascripts/company-edit.bundle.js"></script>
 {% endblock %}
