+++ src/views/company/edit-ukother.html
@@ -113,5 +113,5 @@
       </form>
     </div>
   </div>
-  <script src="/javascripts/companyedit.bundle.js"></script>
+  <script src="/javascripts/company-edit.bundle.js"></script>
 {% endblock %}
