+++ src/views/company/investmentform.html
@@ -80,5 +80,5 @@
     {% endif %}
   </div>
 </form>
-<script src="/javascripts/companyinvestmenttabform.bundle.js"></script>
+<script src="/javascripts/company-investment-tab-form.bundle.js"></script>
 {% endblock %}
