+++ src/views/contact/edit.html
@@ -101,5 +101,5 @@
       </form>
     </div>
   </div>
-  <script src="/javascripts/contactedit.bundle.js"></script>
+  <script src="/javascripts/contact-edit.bundle.js"></script>
 {% endblock %}
