+++ src/views/interaction/servicedelivery-edit.html
@@ -135,5 +135,5 @@
 
     {{ trade.save(backUrl) }}
   </form>
-  <script src="{{ asset_path }}javascripts/servicedelivery.bundle.js"></script>
+  <script src="{{ asset_path }}javascripts/service-delivery.bundle.js"></script>
 {% endblock %}
