+++ src/views/investment/create.html
@@ -143,7 +143,7 @@
     {#</form>#}
   </div>
 
-  <script src="/javascripts/createinvestment.bundle.js"></script>
+  <script src="/javascripts/create-investment.bundle.js"></script>
 
 
 {% endblock %}
